Final Commit

+++ API/DTOs/ImageDto.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace API.DTOs
 {
     public class ImageDto
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string ImageSrc { get; set; }
     }
 }
